Add RouteDispatcher.Tests.Common project and update test project references
